Enhancement: Subplot Support for hist_res Function
Summary:
This update adds support for subplots to the hist_res function, allowing users to specify a subplot position within a larger figure grid. If a subplot argument is provided (as a tuple, e.g., (nrows, ncols, index)), the histogram of residuals will be plotted in the specified location within the subplot grid. If no subplot is provided, a new figure is created.

Changes Made:

Added a subplot argument to hist_res, enabling it to position the histogram and normal distribution curve within a specified subplot.
Modified the function to use plt.subplot(*subplot) if a subplot argument is passed, or plt.figure() to create a new figure when subplot is None.