Enhancement: Subplot Support for plot_res Function
Summary:
This update adds subplot support to the plot_res function, allowing users to specify a subplot location within a grid of plots. If a subplot argument is provided (as a tuple, e.g., (nrows, ncols, index)), the residual plot will be displayed in the specified subplot position. If no subplot is provided, a new figure will be created.

Changes Made:

Added a subplot parameter to plot_res, enabling it to be positioned within a specified subplot grid.
Implemented plt.subplot(*subplot) to place the plot in a grid when subplot is provided, or plt.figure() to generate a new figure when subplot is None.